+++ sql/catalyst/src/main/scala/org/apache/spark/sql/connector/catalog/Catalogs.scala
@@ -40,44 +40,44 @@ private[sql] object Catalogs {
    * @throws CatalogNotFoundException if the plugin class cannot be found
    * @throws SparkException           if the plugin class cannot be instantiated
    */
-    @throws[CatalogNotFoundException]
-    @throws[SparkException]
-    def load(name: String, conf: SQLConf): CatalogPlugin = {
-      var pluginClassName = ""
-      try {
-        pluginClassName = conf.getConfString("spark.sql.catalog." + name)
-      } catch {
-        case _: NoSuchElementException =>
-          throw new CatalogNotFoundException(
-            s"Catalog '$name' plugin class not found: spark.sql.catalog.$name is not defined")
-      }
-      val loader = Utils.getContextOrSparkClassLoader
-      try {
-        val pluginClass = loader.loadClass(pluginClassName)
-        if (!classOf[CatalogPlugin].isAssignableFrom(pluginClass)) throw new SparkException(
-          s"Plugin class for catalog '$name' does not implement CatalogPlugin: $pluginClassName")
-        val plugin = classOf[CatalogPlugin].cast(
-          pluginClass.getDeclaredConstructor().newInstance())
-        plugin.initialize(name, catalogOptions(name, conf))
-        plugin
-      } catch {
-        case _: ClassNotFoundException =>
-          throw new SparkException(
-            s"Cannot find catalog plugin class for catalog '$name': $pluginClassName")
-        case e: NoSuchMethodException =>
-          throw new SparkException(
-            s"Failed to find public no-arg constructor for catalog '$name': $pluginClassName)", e)
-        case e: IllegalAccessException =>
-          throw new SparkException(
-            s"Failed to call public no-arg constructor for catalog '$name': $pluginClassName)", e)
-        case e: InstantiationException =>
-          throw new SparkException("Cannot instantiate abstract catalog plugin class for " +
-            s"catalog '$name': $pluginClassName", e.getCause)
-        case e: InvocationTargetException =>
-          throw new SparkException("Failed during instantiating constructor for catalog " +
-            s"'$name': $pluginClassName", e.getCause)
-      }
+  @throws[CatalogNotFoundException]
+  @throws[SparkException]
+  def load(name: String, conf: SQLConf): CatalogPlugin = {
+    var pluginClassName = ""
+    try {
+      pluginClassName = conf.getConfString("spark.sql.catalog." + name)
+    } catch {
+      case _: NoSuchElementException =>
+        throw new CatalogNotFoundException(
+          s"Catalog '$name' plugin class not found: spark.sql.catalog.$name is not defined")
     }
+    val loader = Utils.getContextOrSparkClassLoader
+    try {
+      val pluginClass = loader.loadClass(pluginClassName)
+      if (!classOf[CatalogPlugin].isAssignableFrom(pluginClass)) throw new SparkException(
+        s"Plugin class for catalog '$name' does not implement CatalogPlugin: $pluginClassName")
+      val plugin = classOf[CatalogPlugin].cast(
+        pluginClass.getDeclaredConstructor().newInstance())
+      plugin.initialize(name, catalogOptions(name, conf))
+      plugin
+    } catch {
+      case _: ClassNotFoundException =>
+        throw new SparkException(
+          s"Cannot find catalog plugin class for catalog '$name': $pluginClassName")
+      case e: NoSuchMethodException =>
+        throw new SparkException(
+          s"Failed to find public no-arg constructor for catalog '$name': $pluginClassName)", e)
+      case e: IllegalAccessException =>
+        throw new SparkException(
+          s"Failed to call public no-arg constructor for catalog '$name': $pluginClassName)", e)
+      case e: InstantiationException =>
+        throw new SparkException("Cannot instantiate abstract catalog plugin class for " +
+          s"catalog '$name': $pluginClassName", e.getCause)
+      case e: InvocationTargetException =>
+        throw new SparkException("Failed during instantiating constructor for catalog " +
+          s"'$name': $pluginClassName", e.getCause)
+    }
+  }
 
   /**
    * Extracts a named catalog's configuration from a SQLConf.
